Make recovery permissive
This is to make it easier for people to build TWRP in the CM tree
if for some reason they desire to do so. Some devices wil not obey
the kernel command line option (exynos, omap) so I am pushing this
patch so that they can easily cherry pick it. 

Author	
Dees Troy <dees_troy@teamw.in>
Feb 21, 2015 10:15 AM
Committer	
Dees Troy <dees_troy@teamw.in>